automl_pyspark/requirements.txt
# Core dependencies
# Note: This requirements.txt is compatible with Python 3.8+ and includes PySpark 3.5.6
# Note: automl_pyspark is installed from local source in Dockerfile
GitPython==3.1.45
# Required for PySpark compatibility with Python 3.13+ (provides distutils)
setuptools>=80.9.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-dataproc==5.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8

# Machine Learning Libraries
lightgbm==4.6.0
xgboost==3.0.3
scikit-learn==1.7.1
scipy==1.16.1
numpy==1.26.4
pandas==2.3.1

# PySpark and Big Data
pyspark==3.5.6
py4j==0.10.9.7
pyarrow==21.0.0
synapseml==1.0.12

# Data Processing and Utilities
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==2.0.1
openpyxl==3.1.5
optuna==4.4.0
packaging==25.0
pillow==11.3.0
plotly==6.2.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
seaborn==0.13.2
shap==0.46.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.42
streamlit==1.47.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.5